third_party/blink/web_tests/virtual/do-not-hide-visited-color/external/wpt/css/css-color/relative-currentcolor-visited-getcomputedstyle-expected.txt

This is a testharness.js-based test.
[FAIL] Property background-color value 'rgb(from currentcolor r g b)' should not leak :visited for computed style
  assert_equals: expected "color(srgb 0 1 0)" but got "color(srgb 1 0 0)"
Harness: the test ran to completion.
